requirements.txt
amqp==2.5.1
anaconda-client==1.2.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
clyent==1.2.1
Django==2.2.3
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-db-geventpool==3.1.0
django-guardian==2.0.0
django-ratelimit==2.0.0
django-redis==4.10.0
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
mysqlclient==1.4.4
Pillow==6.2.0
psycogreen==1.0.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
redis==3.3.10
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
uWSGI==2.0.18
vine==1.3.0
virtualenv==16.7.5
zipp==0.6.0
